Telegram bot two commands

+++ models/house.go
@@ -7,7 +7,7 @@ type House struct {
 	City               string `json:"city"`
 	HouseType          string `json:"house_type,omitempty"`
 	ResidentialComplex string `json:"residential_complex,omitempty"`
-	YearOfBuild        uint16 `json:"year_of_build,omitempty"`
+	YearOfBuild        int    `json:"year_of_build,omitempty"`
 	Floor              string `json:"floor,omitempty"`
 	Area               string `json:"area,omitempty"`
 	Bathroom           string `json:"bathroom,omitempty"`
